unify explorer fullscreen mode and extract column metadata composable
