+++ datagouv-components/src/components/TabularExplorer/TabularExplorer.vue
@@ -72,10 +72,11 @@ import { useFetch } from '../../functions/api'
 import { useComponentsConfig } from '../../config'
 import { useTranslation } from '../../composables/useTranslation'
 import { injectTabularProfile } from '../../composables/useTabularProfile'
-import { buildTypeConfig, buildFormatConfig, humanizeFormat, GENERIC_FORMATS, hasFilterForColumn as _hasFilterForColumn, isTruthy, isFalsy, resolveColumnType } from '../../functions/tabular'
+import { hasFilterForColumn as _hasFilterForColumn } from '../../functions/tabular'
 import SimpleBanner from '../SimpleBanner.vue'
-import type { TabularDataResponse, TabularRow, ColumnType, SortConfig, ColumnFilters, BadgeStyle } from './types'
+import type { TabularDataResponse, TabularRow, SortConfig, ColumnFilters } from './types'
 import { provideTabularContext, type ActiveFilter } from './useTabularContext'
+import { useColumnMetadata } from './useColumnMetadata'
 
 const props = defineProps<{
   resourceId: string
@@ -269,111 +270,20 @@ function hasFilterForColumn(col: string): boolean {
 // Whether the mobile filter sheet is open (button lives in the toolbar).
 const mobileFilterOpen = ref(false)
 
-// Column type helpers
-function getColumnTypeFromName(col: string): ColumnType {
-  const profile = profileData.value?.profile
-  if (!profile) return 'text'
-  const colInfo = profile.columns[col]
-  if (!colInfo) return 'text'
-  return resolveColumnType(colInfo, profile.categorical.includes(col))
-}
-
-const columnTypesMap = computed(() => {
-  const map: Record<string, ColumnType> = {}
-  for (const col of allColumns.value) {
-    map[col] = getColumnTypeFromName(col)
-  }
-  return map
-})
-
-function getColumnType(col: string): ColumnType {
-  return columnTypesMap.value[col] ?? 'text'
-}
-
-function getColumnProfile(col: string) {
-  return profileData.value?.profile?.profile?.[col] ?? null
-}
-
-function getTopsEntries(col: string) {
-  return getColumnProfile(col)?.tops ?? []
-}
-
-function getNullPercent(col: string) {
-  const colProfile = getColumnProfile(col)
-  const total = profileData.value?.profile?.total_lines
-  if (!colProfile || !total) return '0%'
-  return `${((colProfile.nb_missing_values / total) * 100).toFixed(1)}%`
-}
-
-const typeConfig = buildTypeConfig(t)
-const formatConfig = buildFormatConfig(t)
-
-function getTypeConfig(col: string) {
-  return typeConfig[getColumnType(col)]
-}
-
-// Header display: prefer the rich semantic format detected by csv-detective
-// (code INSEE, UAI, département, SIREN, email…) over the generic display type.
-// Tolerant: generic formats fall back to the type config, and unknown semantic
-// formats are humanized rather than dropped.
-function getColumnDisplay(col: string) {
-  const format = profileData.value?.profile?.columns?.[col]?.format
-  const fallback = getTypeConfig(col)
-  if (!format || GENERIC_FORMATS.has(format)) return fallback
-  return formatConfig[format] ?? { icon: fallback.icon, label: humanizeFormat(format) }
-}
-
-const BADGE_PALETTE = [
-  { bg: '#E6EEFE', text: '#3558A2' }, // blue-cumulus
-  { bg: '#E8E3DB', text: '#6A6156' }, // beige-gris-galet
-  { bg: '#FEECC2', text: '#695240' }, // yellow-tournesol
-  { bg: '#C7F6FC', text: '#006A6F' }, // green-archipel
-  { bg: '#E9EDFE', text: '#2323FF' }, // blue-ecume
-  { bg: '#FFE0C7', text: '#885B40' }, // orange-terre-battue
-  { bg: '#F2E9DB', text: '#6B4C35' }, // brown-cafe-creme
-  { bg: '#BAFAEE', text: '#297254' }, // green-menthe
-  { bg: '#FEE0EB', text: '#8D5368' }, // pink-macaron
-  { bg: '#FCE164', text: '#695240' }, // yellow-moutarde
-] as const
-
-const BADGE_FALLBACK = { bg: '#F0E0CF', text: '#745B47' } // brown-opera
-
-const badgeColorMap = computed(() => {
-  const map = new Map<string, { bg: string, text: string }>()
-  const profile = profileData.value?.profile
-  if (!profile) return map
-  for (const col of profile.categorical) {
-    getTopsEntries(col).forEach((top, i) => {
-      map.set(`${col}::${top.value}`, BADGE_PALETTE[i % BADGE_PALETTE.length]!)
-    })
-  }
-  return map
-})
-
-function getCategoryBadgeStyle(col: string, value: string): BadgeStyle {
-  const colors = badgeColorMap.value.get(`${col}::${value}`) ?? BADGE_FALLBACK
-  return { backgroundColor: colors.bg, color: colors.text }
-}
-
-function getCategoryBadgeStylesForColumn(col: string): Record<string, BadgeStyle> {
-  const styles: Record<string, BadgeStyle> = {}
-  for (const top of getTopsEntries(col)) {
-    styles[top.value] = getCategoryBadgeStyle(col, top.value)
-  }
-  return styles
-}
-
-function getBooleanCounts(col: string): { trueCount: number, falseCount: number } {
-  const profile = getColumnProfile(col)
-  if (!profile?.tops) return { trueCount: 0, falseCount: 0 }
-  let trueCount = 0
-  let falseCount = 0
-  for (const top of profile.tops) {
-    if (isTruthy(top.value ?? '')) trueCount += top.count
-    else if (isFalsy(top.value ?? '')) falseCount += top.count
-  }
-  return { trueCount, falseCount }
-}
+// Column metadata (display type, label/icon, null ratio, badge colors, boolean
+// counts) — pure derivations of the profile, extracted to keep this component
+// focused on data fetching, pagination and filter/column state.
+const {
+  columnTypesMap,
+  getColumnType,
+  getColumnProfile,
+  getColumnDisplay,
+  getTopsEntries,
+  getNullPercent,
+  getCategoryBadgeStyle,
+  getCategoryBadgeStylesForColumn,
+  getBooleanCounts,
+} = useColumnMetadata(profileData, allColumns, t)
 
 // Provide the shared state so child components (active filters, columns menu, rows
 // info, table…) can inject it instead of receiving a wall of props.
